refactor tensor and remove duplicate rl
