--- crates/ml/srcs/rl.rs
@@ -1,212 +1,0 @@
-use crate::{nn::{Dense, self}, optim::Adam, recorder::Recorder, tape::Tape, Tensor};
-use std::collections::HashMap;
-
-pub trait Env {
-    fn step(&mut self, action: f32) -> (Vec<f32>, f32, bool);
-    fn reset(&mut self) -> Vec<f32>;
-    fn obs_size(&self) -> usize;
-    fn action_size(&self) -> usize;
-}
-
-pub struct RollingSphereEnv {
-    pos_x: f32,
-}
-
-impl RollingSphereEnv {
-    pub fn new() -> Self {
-        Self { pos_x: 0.0 }
-    }
-}
-
-impl Env for RollingSphereEnv {
-    fn step(&mut self, action: f32) -> (Vec<f32>, f32, bool) {
-        let force = action.max(-10.0).min(10.0);
-        let old_pos_x = self.pos_x;
-        self.pos_x += force * 0.02;
-        let reward = self.pos_x - old_pos_x;
-        let done = self.pos_x.abs() > 5.0;
-        (vec![self.pos_x], reward, done)
-    }
-
-    fn reset(&mut self) -> Vec<f32> {
-        self.pos_x = 0.0;
-        vec![self.pos_x]
-    }
-
-    fn obs_size(&self) -> usize { 1 }
-    fn action_size(&self) -> usize { 1 }
-}
-
-pub struct PolicyValueNet {
-    l1: Dense,
-    policy_head: Dense,
-    value_head: Dense,
-}
-
-impl PolicyValueNet {
-    pub fn new(in_dim: usize, hidden_dim: usize, out_dim: usize) -> Self {
-        Self {
-            l1: Dense::random(in_dim, hidden_dim, 0),
-            policy_head: Dense::random(hidden_dim, out_dim, 1),
-            value_head: Dense::random(hidden_dim, 1, 2),
-        }
-    }
-
-    pub fn forward(&self, x: &Tensor, recorder: &mut impl Recorder) -> (Tensor, Tensor) {
-        let (x, _) = self.l1.forward(x, recorder);
-        let x = x.tanh(recorder);
-        let (policy, _) = self.policy_head.forward(&x, recorder);
-        let (value, _) = self.value_head.forward(&x, recorder);
-        (policy, value)
-    }
-
-    pub fn get_params(&mut self) -> Vec<&mut Tensor> {
-        vec![&mut self.l1.w, &mut self.l1.b, &mut self.policy_head.w, &mut self.policy_head.b, &mut self.value_head.w, &mut self.value_head.b]
-    }
-}
-
-pub struct SpherePpoTrainer {
-    envs: Vec<RollingSphereEnv>,
-    net: PolicyValueNet,
-    optimizer: Adam,
-    gamma: f32,
-    lambda: f32,
-    clip: f32,
-    t_max: usize,
-    n_epochs: usize,
-    obs: Vec<Vec<f32>>,
-}
-
-impl SpherePpoTrainer {
-    pub fn new(seed: u64) -> Self {
-        fastrand::seed(seed);
-        let mut net = PolicyValueNet::new(1, 32, 1);
-        let optimizer = Adam::new(&net.get_params().iter().map(|t| &**t).collect::<Vec<&Tensor>>());
-        let obs = (0..8).map(|_| vec![0.0]).collect();
-        Self {
-            envs: (0..8).map(|_| RollingSphereEnv::new()).collect(),
-            net,
-            optimizer,
-            gamma: 0.99,
-            lambda: 0.95,
-            clip: 0.2,
-            t_max: 64,
-            n_epochs: 4,
-            obs,
-        }
-    }
-
-    pub fn step(&mut self) -> f32 {
-        let mut all_obs = Vec::new();
-        let mut all_actions = Vec::new();
-        let mut all_log_probs = Vec::new();
-        let mut all_rewards = Vec::new();
-        let mut all_dones = Vec::new();
-        let mut all_values = Vec::new();
-        let mut total_rewards = vec![0.0; self.envs.len()];
-
-        for _ in 0..self.t_max {
-            all_obs.push(self.obs.clone());
-            let obs_tensor = Tensor::from_vec(vec![self.envs.len(), 1], self.obs.iter().flatten().copied().collect());
-            let (action_tensor, value_tensor) = self.net.forward(&obs_tensor, &mut nn::graph::Graph::new());
-            
-            let actions = action_tensor.data().to_vec();
-            all_actions.push(actions.clone());
-            all_values.push(value_tensor.data().to_vec());
-            all_log_probs.push(actions.iter().map(|a| -0.5 * a.powi(2) - 0.5 * (2.0 * std::f32::consts::PI).ln()).collect::<Vec<_>>());
-
-            let mut next_obs_vec = Vec::new();
-            let mut rewards = Vec::new();
-            let mut dones = Vec::new();
-
-            for (i, env) in self.envs.iter_mut().enumerate() {
-                let (next_ob, reward, done) = env.step(actions[i]);
-                next_obs_vec.push(next_ob);
-                rewards.push(reward);
-                dones.push(done);
-                total_rewards[i] += reward;
-                if done {
-                    self.obs[i] = env.reset();
-                } else {
-                    self.obs[i] = next_obs_vec[i].clone();
-                }
-            }
-            all_rewards.push(rewards);
-            all_dones.push(dones);
-        }
-
-        let mut advantages = vec![vec![0.0; self.envs.len()]; self.t_max];
-        let mut returns = vec![vec![0.0; self.envs.len()]; self.t_max];
-        let mut last_advantage = vec![0.0; self.envs.len()];
-
-        let obs_tensor = Tensor::from_vec(vec![self.envs.len(), 1], self.obs.iter().flatten().copied().collect());
-        let (_, last_values_tensor) = self.net.forward(&obs_tensor, &mut nn::graph::Graph::new());
-        let last_values = last_values_tensor.data();
-
-        for t in (0..self.t_max).rev() {
-            for i in 0..self.envs.len() {
-                let next_value = if t == self.t_max - 1 { last_values[i] } else { all_values[t + 1][i] };
-                let next_is_done = if t == self.t_max - 1 { false } else { all_dones[t+1][i] };
-                let delta = all_rewards[t][i] + self.gamma * next_value * (1.0 - next_is_done as i32 as f32) - all_values[t][i];
-                advantages[t][i] = delta + self.gamma * self.lambda * last_advantage[i] * (1.0 - next_is_done as i32 as f32);
-                last_advantage[i] = advantages[t][i];
-                returns[t][i] = advantages[t][i] + all_values[t][i];
-            }
-        }
-        
-        let obs_flat: Vec<f32> = all_obs.into_iter().flatten().flatten().collect();
-        let actions_flat: Vec<f32> = all_actions.into_iter().flatten().collect();
-        let log_probs_flat: Vec<f32> = all_log_probs.into_iter().flatten().collect();
-        let advantages_flat: Vec<f32> = advantages.into_iter().flatten().collect();
-        let returns_flat: Vec<f32> = returns.into_iter().flatten().collect();
-
-        let advantages_mean = advantages_flat.iter().sum::<f32>() / advantages_flat.len() as f32;
-        let advantages_std = (advantages_flat.iter().map(|x| (x - advantages_mean).powi(2)).sum::<f32>() / advantages_flat.len() as f32).sqrt();
-        let advantages_norm: Vec<f32> = advantages_flat.iter().map(|x| (x - advantages_mean) / (advantages_std + 1e-8)).collect();
-        
-        for _ in 0..self.n_epochs {
-            let mut tape = Tape::new();
-            let mut tensors = HashMap::new();
-            
-            for p in self.net.get_params().iter_mut() {
-                p.set_requires_grad();
-                tensors.insert(p.id, (*p).clone());
-            }
-
-            let obs_tensor = Tensor::from_vec(vec![obs_flat.len(), 1], obs_flat.clone());
-            let (action_tensor, value_tensor) = self.net.forward(&obs_tensor, &mut tape);
-            
-            let log_probs = action_tensor.sub(&Tensor::from_vec(vec![actions_flat.len()], actions_flat.clone()), &mut tape).pow(2.0, &mut tape).mul_scalar(-0.5, &mut tape);
-            let advantages_tensor = Tensor::from_vec(vec![advantages_norm.len()], advantages_norm.clone());
-            let returns_tensor = Tensor::from_vec(vec![returns_flat.len()], returns_flat.clone());
-            let old_log_probs_tensor = Tensor::from_vec(vec![log_probs_flat.len()], log_probs_flat.clone());
-
-            let ratio = log_probs.sub(&old_log_probs_tensor, &mut tape).exp(&mut tape);
-            
-            let policy_loss1 = ratio.mul(&advantages_tensor, &mut tape);
-            let policy_loss2 = ratio.clone().clamp(1.0 - self.clip, 1.0 + self.clip, &mut tape).mul(&advantages_tensor, &mut tape);
-            let policy_loss = policy_loss1.min(&policy_loss2, &mut tape).reduce_mean(&mut tape).mul_scalar(-1.0, &mut tape);
-
-            let value_loss = value_tensor.sub(&returns_tensor, &mut tape).pow(2.0, &mut tape).reduce_mean(&mut tape);
-            
-            let loss = policy_loss.add(&value_loss, &mut tape);
-            
-            let intermediate_tensors = tape.nodes().iter().map(|n| n.out).collect::<Vec<_>>();
-            for id in intermediate_tensors {
-                // This is a hack to get the intermediate tensors into the map. A better solution
-                // would be to have the recorder populate the map directly.
-                if !tensors.contains_key(&id) {
-                    // This is very inefficient, but it will work for the test.
-                    // A real implementation would need a way to get a tensor by its ID.
-                }
-            }
-            tensors.insert(loss.id, loss.clone());
-
-            tape.backward(&loss, &mut tensors).unwrap();
-
-            self.optimizer.step(&mut self.net.get_params());
-        }
-
-        total_rewards.iter().sum::<f32>() / total_rewards.len() as f32
-    }
-}
